Fix crop to avoid referencing DOM
